shake: fix diagrams
Previously the diagrams would not get rebuilt after deleting `_build`,
because the oracle rule was cached. Fix this by properly tracking the
generated PDF and SVG files.

This is a bit awkward because we want to specify the rule forwards
since the SVG filename depends on the contents, but Shake works
backwards. The solution implemented here is to write the contents to a
TeX file (using writeFileChanged to avoid triggering unnecessary
rebuilds), then `need` the final SVG file, which is built using normal
"backwards" rules.